+++ TextRPG2/Goblin.h
@@ -12,12 +12,12 @@ private:
 	virtual ~Goblin() = default;
 
 public: 
-	virtual void Initialize(string name, int health, int attack) override;
+	virtual void Initialize(string name, double health, double attack) override;
 	virtual void Update() override;
 	virtual void Interact() override;
 
 public: /* 실제 동적할당을 대신해주는 함수 */
-	static Goblin *Create(string name = "", int health = 10, int attack = 5);
-	static Goblin* CreateBoss(string name, int health, int attack);
+	static Goblin *Create(string name = "", double health = 10, double attack = 5);
+	static Goblin* CreateBoss(string name, double health, double attack);
 	virtual void Free() override;
 };
