+++ TextRPG2/Orc.cpp
@@ -5,7 +5,7 @@ Orc::Orc()
 {
 }
 
-void Orc::Initialize(string name, int health, int attack)
+void Orc::Initialize(string name, double health, double attack)
 {
 	mName = name;
 	mHealth = health;
@@ -21,7 +21,7 @@ void Orc::Interact() {
 
 }
 
-Orc* Orc::Create(string name, int health, int attack)
+Orc* Orc::Create(string name, double health, double attack)
 {
 	Orc* pOrc = new Orc();
 
@@ -30,7 +30,7 @@ Orc* Orc::Create(string name, int health, int attack)
 	return pOrc;
 }
 
-Orc* Orc::CreateBoss(string name, int health, int attack)
+Orc* Orc::CreateBoss(string name, double health, double attack)
 {
 	Orc* pOrc = new Orc();
