+++ TextRPG2/Slime.cpp
@@ -5,7 +5,7 @@ Slime::Slime()
 {
 }
 
-void Slime::Initialize(string name, int health, int attack)
+void Slime::Initialize(string name, double health, double attack)
 {
 	mName = name;
 	mHealth = health;
@@ -21,7 +21,7 @@ void Slime::Interact() {
 
 }
 
-Slime* Slime::Create(string name, int health, int attack)
+Slime* Slime::Create(string name, double health, double attack)
 {
 	Slime* pSlime = new Slime();
 
@@ -30,7 +30,7 @@ Slime* Slime::Create(string name, int health, int attack)
 	return pSlime;
 }
 
-Slime* Slime::CreateBoss(string name, int health, int attack)
+Slime* Slime::CreateBoss(string name, double health, double attack)
 {
 	Slime* pSlime = new Slime();
