+++ TextRPG2/Slime.h
@@ -9,12 +9,12 @@ private:
 	virtual ~Slime() = default;
 
 public:
-	virtual void Initialize(string name, int health, int attack) override;
+	virtual void Initialize(string name, double health, double attack) override;
 	virtual void Update() override;
 	virtual void Interact() override;
 
 public:
-	static Slime* Create(string name, int health, int attack);
-	static Slime* CreateBoss(string name, int health, int attack);
+	static Slime* Create(string name, double health, double attack);
+	static Slime* CreateBoss(string name, double health, double attack);
 	virtual void Free() override;
 };
